refactor: #9 新規メモ作成処理をリファクタリング。
