+++ bear-clone/src/app/shared/actions/note.actions.ts
@@ -5,7 +5,8 @@ import { ContentChange } from "ngx-quill";
 export const NoteActions = createActionGroup({
   source: "Note Local Storage",
   events: {
-    "Add Notes": props<{ newNotes: Note[] }>(),
+    "Add Notes": emptyProps(),
+    "Add Notes Success": props<{ newNotes: Note[] }>(),
     "Remove Notes": props<{ newNotes: Note[] }>(),
     "Update Notes": props<{ newNotes: Note[] }>(),
   },
